feat(member-card): add MemberCard component
Add a new component called MemberCard to the project. This component is a card that displays information about a member. It receives three props: username, image, and birthday. The username and image are used to display the member's name and profile picture, respectively. The birthday prop is optional and is used to display the member's age. The component is exported from the index file and the type definition is added to the member-card.type.ts file.